Fix for issue #22965
